# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Pet Store API",
    "description": "API for managing pets"
  },
  "host": "localhost:3500",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
    "/users": {
      "post": {
        "summary": "Create User",
        "description": "Create a new user",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User to create",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "type": "string" },
                "lastName": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" },
                "allowExtraEmails": { "type": "boolean" }
              },
              "required": ["firstName", "lastName", "email", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "type": "object",
              "properties": {
                "token": { "type": "string" },
                "user": {
                  "type": "object",
                  "properties": {
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" },
                    "allowExtraEmails": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login User",
        "description": "Login a user",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User credentials",
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "token": { "type": "string" },
                "user": {
                  "type": "object",
                  "properties": {
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid email or password" }
        }
      }
    },
    "/pets": {
      "post": {
        "summary": "Create Pet",
        "description": "Create a new pet",
        "parameters": [
          {
            "in": "body",
            "name": "pet",
            "description": "Pet to create",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "type": { "type": "string" },
                "breed": { "type": "string" },
                "price": { "type": "number" }
              },
              "required": ["name", "type", "breed", "price"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Pet created",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "number" },
                "name": { "type": "string" },
                "type": { "type": "string" },
                "breed": { "type": "string" },
                "price": { "type": "number" }
              }
            }
          },
          "400": { "description": "Invalid input" }
        }
      },
      "get": {
        "summary": "Get All Pets",
        "description": "Get all pets",
        "responses": {
          "200": {
            "description": "A list of pets",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "number" },
                  "name": { "type": "string" },
                  "type": { "type": "string" },
                  "breed": { "type": "string" },
                  "price": { "type": "number" }
                }
              }
            }
          }
        }
      }
    },
    "/pets/{id}": {
      "put": {
        "summary": "Update Pet",
        "description": "Update a pet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "pet",
            "description": "Pet to update",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "type": { "type": "string" },
                "breed": { "type": "string" },
                "price": { "type": "number" }
              },
              "required": ["name", "type", "breed", "price"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet updated",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "number" },
                "name": { "type": "string" },
                "type": { "type": "string" },
                "breed": { "type": "string" },
                "price": { "type": "number" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "404": { "description": "Pet not found" }
        }
      },
      "delete": {
        "summary": "Delete Pet",
        "description": "Delete a pet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Pet deleted",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "number" },
                "name": { "type": "string" },
                "type": { "type": "string" },
                "breed": { "type": "string" },
                "price": { "type": "number" }
              }
            }
          },
          "404": { "description": "Pet not found" }
        }
      }
    }
  }
}